RASSILKA_BOT1/requirements.txt
aiogram>=3.26,<4.0
python-dotenv>=1.0,<2.0